+++ pkg/main.go
@@ -1,27 +1,16 @@
 package main
 
 import (
-	"net/http"
 	"os"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
+	"github.com/grafana/google-sheets-datasource/pkg/googlesheets"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
-	"github.com/grafana/grafana-plugin-sdk-go/backend/resource/httpadapter"
-	"github.com/grafana/grafana-plugin-sdk-go/experimental"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 )
 
 func main() {
-	mux := http.NewServeMux()
-	ds := NewDataSource(mux)
-	httpResourceHandler := httpadapter.New(mux)
-
-	err := experimental.DoGRPC("google-sheets-datasource", datasource.ServeOpts{
-		CallResourceHandler: httpResourceHandler,
-		QueryDataHandler:    ds,
-		CheckHealthHandler:  ds,
-	})
-	if err != nil {
-		backend.Logger.Error(err.Error())
+	if err := datasource.Manage("google-sheets-datasource", googlesheets.NewDatasource, datasource.ManageOpts{}); err != nil {
+		log.DefaultLogger.Error(err.Error())
 		os.Exit(1)
 	}
 }