+++ src/components/ConfigEditor.tsx
@@ -1,179 +1,50 @@
-import React, { PureComponent } from 'react';
-import { LegacyForms, Select, InlineFormLabel } from '@grafana/ui';
-import {
-  DataSourcePluginOptionsEditorProps,
-  onUpdateDatasourceSecureJsonDataOption,
-  onUpdateDatasourceJsonDataOptionSelect,
-} from '@grafana/data';
-import { SheetsSourceOptions, GoogleSheetsSecureJsonData, GoogleAuthType, googleAuthTypes } from '../types';
-import { JWTConfig } from './';
+import { DataSourcePluginOptionsEditorProps, onUpdateDatasourceSecureJsonDataOption } from '@grafana/data';
+import { AuthConfig, DataSourceOptions } from '@grafana/google-sdk';
+import { Field, LegacyForms, SecretInput } from '@grafana/ui';
+import React from 'react';
+import { GoogleSheetsAuth, GoogleSheetsSecureJSONData, googleSheetsAuthTypes } from '../types';
+import { getBackwardCompatibleOptions } from '../utils';
+import { ConfigurationHelp } from './ConfigurationHelp';
 
-export type Props = DataSourcePluginOptionsEditorProps<SheetsSourceOptions, GoogleSheetsSecureJsonData>;
+const { SecretFormField } = LegacyForms;
 
-export class ConfigEditor extends PureComponent<Props> {
-  onResetApiKey = () => {
-    const { options } = this.props;
-    this.props.onOptionsChange({
-      ...options,
-      secureJsonData: {
-        ...options.secureJsonData,
-        apiKey: '',
-      },
-      secureJsonFields: {
-        ...options.secureJsonFields,
-        apiKey: false,
-      },
-    });
+export type Props = DataSourcePluginOptionsEditorProps<DataSourceOptions, GoogleSheetsSecureJSONData>;
+
+export function ConfigEditor(props: Props) {
+  const options = getBackwardCompatibleOptions(props.options);
+  const apiKeyProps = {
+    isConfigured: Boolean(options.secureJsonFields.apiKey),
+    value: options.secureJsonData?.apiKey || '',
+    placeholder: 'Enter API key',
+    id: 'apiKey',
+    onReset: () =>
+      props.onOptionsChange({
+        ...options,
+        secureJsonFields: { ...options.secureJsonFields, apiKey: false },
+        secureJsonData: { apiKey: '' },
+        jsonData: options.jsonData,
+      }),
+    onChange: onUpdateDatasourceSecureJsonDataOption(props, 'apiKey'),
   };
 
-  render() {
-    const { options, onOptionsChange } = this.props;
-    const { secureJsonFields, jsonData } = options;
+  return (
+    <>
+      <ConfigurationHelp authenticationType={options.jsonData.authenticationType} />
 
-    if (!jsonData.hasOwnProperty('authType')) {
-      jsonData.authType = GoogleAuthType.KEY;
-    }
+      <AuthConfig authOptions={googleSheetsAuthTypes} onOptionsChange={props.onOptionsChange} options={options} />
 
-    const secureJsonData = options.secureJsonData as GoogleSheetsSecureJsonData;
-    return (
-      <div className="gf-form-group">
-        <div className="gf-form">
-          <InlineFormLabel
-            className="width-10"
-            tooltip="API Key auth is used to access public spreadsheets, and Google JWT File auth using a service account is used to access private files."
-          >
-            Auth
-          </InlineFormLabel>
-          <Select
-            className="width-30"
-            value={googleAuthTypes.find((x) => x.value === jsonData.authType) || googleAuthTypes[0]}
-            options={googleAuthTypes}
-            defaultValue={jsonData.authType}
-            onChange={onUpdateDatasourceJsonDataOptionSelect(this.props, 'authType')}
-          />
-        </div>
-        {jsonData.authType === GoogleAuthType.KEY && (
-          <>
-            <div className="gf-form">
-              <LegacyForms.SecretFormField
-                isConfigured={(secureJsonFields && secureJsonFields.apiKey) as boolean}
-                value={secureJsonData?.apiKey || ''}
-                label="API Key"
-                labelWidth={10}
-                inputWidth={30}
-                placeholder="Enter API Key"
-                onReset={this.onResetApiKey}
-                onChange={onUpdateDatasourceSecureJsonDataOption(this.props, 'apiKey')}
-              />
-            </div>
-          </>
-        )}
-        {jsonData.authType === GoogleAuthType.JWT && (
-          <JWTConfig
-            isConfigured={(secureJsonFields && !!secureJsonFields.jwt) as boolean}
-            onChange={(jwt) => {
-              onOptionsChange({
-                ...options,
-                secureJsonData: {
-                  ...secureJsonData,
-                  jwt,
-                },
-              });
-            }}
-          ></JWTConfig>
-        )}
-        <div className="grafana-info-box" style={{ marginTop: 24 }}>
-          {jsonData.authType === GoogleAuthType.JWT ? (
-            <>
-              <h4>Generate a JWT file</h4>
-              <ol style={{ listStylePosition: 'inside' }}>
-                <li>
-                  Open the{' '}
-                  <a
-                    href="https://console.developers.google.com/apis/credentials"
-                    target="_blank"
-                    rel="noreferrer noopener"
-                  >
-                    Credentials
-                  </a>{' '}
-                  page in the Google API Console.
-                </li>
-                <li>
-                  Click <strong>Create Credentials</strong> then click <strong>Service account</strong>.
-                </li>
-                <li>On the Create service account page, enter the Service account details.</li>
-                <li>
-                  On the <code>Create service account</code> page, fill in the <code>Service account details</code> and
-                  then click <code>Create</code>
-                </li>
-                <li>
-                  On the <code>Service account permissions</code> page, don&rsquo;t add a role to the service account.
-                  Just click <code>Continue</code>
-                </li>
-                <li>
-                  In the next step, click <code>Create Key</code>. Choose key type <code>JSON</code> and click{' '}
-                  <code>Create</code>. A JSON key file will be created and downloaded to your computer
-                </li>
-                <li>
-                  Open the{' '}
-                  <a
-                    href="https://console.cloud.google.com/apis/library/sheets.googleapis.com?q=sheet"
-                    target="_blank"
-                    rel="noreferrer noopener"
-                  >
-                    Google Sheets
-                  </a>{' '}
-                  in API Library and enable access for your account
-                </li>
-                <li>
-                  Open the{' '}
-                  <a
-                    href="https://console.cloud.google.com/apis/library/drive.googleapis.com?q=drive"
-                    target="_blank"
-                    rel="noreferrer noopener"
-                  >
-                    Google Drive
-                  </a>{' '}
-                  in API Library and enable access for your account. Access to the Google Drive API is used to list all
-                  spreadsheets that you have access to.
-                </li>
-                <li>
-                  Share any private files/folders you want to access with the service account&apos;s email address. The
-                  email is specified as <code>client_email</code> in the Google JWT File.
-                </li>
-                <li>
-                  Drag the file to the dotted zone above. Then click <code>Save & Test</code>. The file contents will be
-                  encrypted and saved in the Grafana database.
-                </li>
-              </ol>
-            </>
+      {options.jsonData.authenticationType === GoogleSheetsAuth.API && (
+        <>
+          {/* Backward compatibility check. SecretInput was added in 8.5 */}
+          {!!SecretInput ? (
+            <Field label="API key">
+              <SecretInput {...apiKeyProps} width={60} />
+            </Field>
           ) : (
-            <>
-              <h4>Generate an API key</h4>
-              <ol style={{ listStylePosition: 'inside' }}>
-                <li>
-                  Open the{' '}
-                  <a
-                    href="https://console.developers.google.com/apis/credentials"
-                    target="_blank"
-                    rel="noreferrer noopener"
-                  >
-                    Credentials page
-                  </a>{' '}
-                  in the Google API Console.
-                </li>
-                <li>
-                  Click <strong>Create Credentials</strong> and then click <strong>API key</strong>.
-                </li>
-                <li>
-                  Copy the key and paste it in the API Key field above. The file contents are encrypted and saved in the
-                  Grafana database.
-                </li>
-              </ol>
-            </>
+            <SecretFormField {...apiKeyProps} label="API key" labelWidth={10} inputWidth={20} />
           )}
-        </div>
-      </div>
-    );
-  }
+        </>
+      )}
+    </>
+  );
 }