+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## v1.2.0
+
+- Refactored authentication to use grafana google sdks. With this change you can now use GCE authentication with google sheets.
+
+There was a change in the plugin configuration. Please take a look at the provisioning example in the [documentation](src/docs/provisioning.md).
+The change is backward compatible so you can still use the old configuration.
+
 ## v1.1.8
+
 - **Chore**: Backend binaries are now compiled with golang 1.20.4
 
 ## v1.1.7
+
 - **Chore**: Update to Golang 1.20
 
 ## v1.1.6